delete : delete variable db in app dir
